feat: simplified public api for creating schema and loading settings
All the imports are now done from the root of the package. This makes it
easier to use the package.

Instead of importing the Settings class, now the user can import the
default object with a `schema` function. This function returns a
SettingsBuilder object that can be used to create a schema and load
configuration.